Change all the packages from openecomp to onap
Issue-ID: AAI-61 AAI-82

+++ aai-traversal/src/main/resources/docker/init-chef.sh
@@ -1,24 +1,25 @@
 #!/bin/bash
-
-###
+#
 # ============LICENSE_START=======================================================
-# org.openecomp.aai
+# org.onap.aai
 # ================================================================================
-# Copyright (C) 2017 AT&T Intellectual Property. All rights reserved.
+# Copyright © 2017 AT&T Intellectual Property. All rights reserved.
 # ================================================================================
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-# 
-#      http://www.apache.org/licenses/LICENSE-2.0
-# 
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ============LICENSE_END=========================================================
-###
+#
+# ECOMP is a trademark and service mark of AT&T Intellectual Property.
+#
 
 ##############################################################################
 #       Script to initialize the chef-repo branch and.chef
@@ -34,10 +35,15 @@ if [ ! -d "aai-config" ]; then
         exit;
     }
 
+fi
+
+if [ -d "aai-config/cookbooks/aai-traversal" ]; then
+
     (cd aai-config/cookbooks/aai-traversal/ && \
         for f in $(ls); do mv $f ../; done && \
         cd ../ && rmdir aai-traversal);
-fi
+
+fi;
 
 if [ ! -d "aai-data" ]; then
 
@@ -46,9 +52,10 @@ if [ ! -d "aai-data" ]; then
         exit;
     }
 
-    chef-solo \
-       -c /var/chef/aai-data/chef-config/dev/.knife/solo.rb \
-       -j /var/chef/aai-config/cookbooks/runlist-aai-traversal.json \
-       -E ${AAI_CHEF_ENV};
-
 fi
+
+chef-solo \
+   -c /var/chef/aai-data/chef-config/dev/.knife/solo.rb \
+   -j /var/chef/aai-config/cookbooks/runlist-aai-traversal.json \
+   -E ${AAI_CHEF_ENV};
+
